Move texture decoder
